requirements.txt
# MCP requirements
mcp>=0.0.11

# Google API requirements
google-auth>=2.25.2
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0

# Google Ads API library
google-ads>=22.0.0

# Environment configuration
python-dotenv>=1.0.0

# Optional visualization dependencies
matplotlib>=3.7.3
pandas>=2.1.4
fastapi
uvicorn[standard]

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
pytest-html>=3.2.0
httpx>=0.24.1
responses>=0.23.3